requirements.txt
aiohttp>=3.9.0
aiohttp-jinja2>=1.6.0
jinja2>=3.1.2
cachetools>=5.3.0
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.7.0
onnxruntime>=1.18.0
PyYAML>=6.0.0
pycryptodome
ujson